+++ src/app/services/wallet.service.spec.ts
@@ -34,6 +34,25 @@ describe('WalletService', () => {
     httpTestingController.verify();
   });
 
+  it('should execute content and return a JSON response', (done) => {
+    const mockUrl = 'https://example.com/mock-content';
+    const mockResponse = { success: true, message: 'Content executed successfully' };
+  
+    service.executeContent(mockUrl).subscribe((response) => {
+      expect(response).toEqual(mockResponse);
+      done();
+    });
+  
+    const req = httpTestingController.expectOne(
+      `${environment.server_url}${environment.server_uri.execute_content_uri}`
+    );
+    expect(req.request.method).toEqual('POST');
+    expect(req.request.body).toEqual({ qr_content: mockUrl });
+    expect(req.request.headers.get('Content-Type')).toBe('application/json'); // Opcional: comprovar headers si cal
+    req.flush(mockResponse);
+  });
+  
+
   it('should fetch VC in CBOR format', (done) => {
     const mockCredential: VerifiableCredential = {
       '@context': ['https://www.w3.org/ns/credentials/v1'],