+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [v1.4.0](https://github.com/in2workspace/in2-wallet-ui/releases/tag/v1.4.0)
 ### Added
 - New route to execute same-device credential issuance workflow.
+- Timeout counter added to "Enter PIN" popup
+- More informative messages in case of error in the process of sending PIN to get credential
 
 ## [v1.3.7](https://github.com/in2workspace/in2-wallet-ui/releases/tag/v1.3.7)
 ### Fixed